# Conflicts:
#	impndday.f
#	main.f
#	modparm.f
#	potholehr.f
#	readfile.f
#	rseday.f
#	rsedmon.f
#	rsedyr.f
#	writem.f

+++ pestw.f
@@ -65,7 +65,7 @@
       use parm
 
       integer :: k, kk, j, l
-      real :: pfp, pfg
+      real*8 :: pfp, pfg
 
       call outprocess("pestw")
